# web/openapi/faultevolve.openapi.json
{
  "components": {
    "schemas": {
...
      "DeploymentMode": {
        "description": "How the engine gets executed. P0 ships cloud-only modes.",
        "enum": [
          "local_cloud",
          "ssh_cloud",
          "local_desktop"
        ],
        "title": "DeploymentMode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PathProbeRequest": {
        "additionalProperties": false,
        "description": "Body of a directory check on an execution target.\n\nThe path is checked against ``WebSettings``' allow-list before anything is\nstat'ed, so ``..`` traversal is rejected rather than normalised away.",
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "PathProbeRequest",
        "type": "object"
      },
      "PathProbeResponse": {
        "additionalProperties": false,
        "description": "Result of checking one directory on an execution target.",
        "properties": {
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "exists": {
            "title": "Exists",
            "type": "boolean"
          },
          "is_directory": {
            "title": "Is Directory",
            "type": "boolean"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "writable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Writable"
          }
        },
        "required": [
          "path",
          "exists",
          "is_directory"
        ],
        "title": "PathProbeResponse",
        "type": "object"
      },
...
      "ServerProbeResponse": {
        "additionalProperties": false,
        "description": "Reachability + environment report for one execution target.",
        "properties": {
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "engine_installed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Engine Installed"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Engine Version"
          },
          "gpu_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gpu Summary"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Latency Ms"
          },
          "llm_key_present": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Llm Key Present"
          },
          "mode": {
            "$ref": "#/components/schemas/DeploymentMode"
          },
          "paths": {
            "items": {
              "$ref": "#/components/schemas/PathProbeResponse"
            },
            "title": "Paths",
            "type": "array"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "python_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Python Version"
          },
          "reachable": {
            "title": "Reachable",
            "type": "boolean"
          }
        },
        "required": [
          "profile_id",
          "mode",
          "reachable"
        ],
        "title": "ServerProbeResponse",
        "type": "object"
      },
      "ServerProfileListResponse": {
        "additionalProperties": false,
        "description": "Every stored execution target, plus which one the UI should preselect.",
        "properties": {
          "default_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Default Profile Id"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ServerProfileResponse"
            },
            "title": "Profiles",
            "type": "array"
          },
          "ssh_available": {
            "default": false,
            "title": "Ssh Available",
            "type": "boolean"
          }
        },
        "title": "ServerProfileListResponse",
        "type": "object"
      },
      "ServerProfileRequest": {
        "additionalProperties": false,
        "description": "Create/update a cloud execution target.\n\nSecrets are never part of the request: the client sends a *reference*\n(``private_key_ref``) and the server resolves it locally.",
        "properties": {
          "artifacts_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Artifacts Dir"
          },
          "conda_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Conda Env"
          },
          "cpu_limit": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cpu Limit"
          },
          "data_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Data Dir"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "gpu_count": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gpu Count"
          },
          "gpu_devices": {
            "items": {
              "type": "string"
            },
            "title": "Gpu Devices",
            "type": "array"
          },
          "gpu_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gpu Model"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Host"
          },
          "logs_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Logs Dir"
          },
          "max_concurrency": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Max Concurrency"
          },
          "memory_gb": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Memory Gb"
          },
          "mode": {
            "$ref": "#/components/schemas/DeploymentMode"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "maximum": 65535,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Port"
          },
          "private_key_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Private Key Ref"
          },
          "python_executable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Python Executable"
          },
          "repo_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Repo Dir"
          },
          "runs_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Runs Dir"
          },
          "timeout_s": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Timeout S"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "User"
          }
        },
        "required": [
          "display_name",
          "mode"
        ],
        "title": "ServerProfileRequest",
        "type": "object"
      },
      "ServerProfileResponse": {
        "additionalProperties": false,
        "description": "A stored profile, with every secret replaced by a masked marker.",
        "properties": {
          "artifacts_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Artifacts Dir"
          },
          "conda_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Conda Env"
          },
          "cpu_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Cpu Limit"
          },
          "data_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Data Dir"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "gpu_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gpu Count"
          },
          "gpu_devices": {
            "items": {
              "type": "string"
            },
            "title": "Gpu Devices",
            "type": "array"
          },
          "gpu_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gpu Model"
          },
          "has_private_key": {
            "title": "Has Private Key",
            "type": "boolean"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Host"
          },
          "logs_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Logs Dir"
          },
          "max_concurrency": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Max Concurrency"
          },
          "memory_gb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Memory Gb"
          },
          "mode": {
            "$ref": "#/components/schemas/DeploymentMode"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Port"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "python_executable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Python Executable"
          },
          "repo_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Repo Dir"
          },
          "runs_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Runs Dir"
          },
          "secret_masked": {
            "title": "Secret Masked",
            "type": "boolean"
          },
          "timeout_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Timeout S"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "User"
          }
        },
        "required": [
          "profile_id",
          "display_name",
          "mode",
          "has_private_key",
          "secret_masked"
        ],
        "title": "ServerProfileResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only control and replay surface for the FaultEvolve engine. Scores are produced by the task evaluator and only ever read back here.",
    "title": "FaultEvolve Web API",
    "version": "1.11.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/servers": {
      "get": {
        "description": "Masked profiles plus the default one to preselect (plan D-E).",
        "operationId": "list_servers_api_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerProfileListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "列出服务器配置",
        "tags": [
          "servers"
        ]
      },
      "post": {
        "description": "Secrets never travel: only a ``private_key_ref`` is accepted.",
        "operationId": "upsert_server_api_servers_post",
        "parameters": [
          {
            "in": "query",
            "name": "profile_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "新建或更新服务器配置",
        "tags": [
          "servers"
        ]
      }
    },
    "/api/servers/{profile_id}": {
      "delete": {
        "operationId": "delete_server_api_servers__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "删除服务器配置",
        "tags": [
          "servers"
        ]
      },
      "get": {
        "operationId": "get_server_api_servers__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取单个服务器配置",
        "tags": [
          "servers"
        ]
      }
    },
    "/api/servers/{profile_id}/check-path": {
      "post": {
        "description": "Existence / writability / free space for one directory.\n\nThe path is checked against the allow-list before anything is\nstat'ed, locally and remotely alike.",
        "operationId": "check_path_api_servers__profile_id__check_path_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathProbeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "测试目录",
        "tags": [
          "servers"
        ]
      }
    },
    "/api/servers/{profile_id}/probe": {
      "post": {
        "description": "Reachability + environment report (PRD 7.4 交互规则 2).",
        "operationId": "probe_server_api_servers__profile_id__probe_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerProbeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "测试连接",
        "tags": [
          "servers"
        ]
      }
    },
...
    "/api/tasks/{task_id}": {
      "get": {
        "description": "Sanitised problem/prompt documents plus file and data listings.",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取单个任务的详情",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{task_id}/card": {
      "get": {
        "description": "The overview card (PRD 7.7): one request, no client-side joins.",
        "operationId": "get_task_card_api_tasks__task_id__card_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDatasetCardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "读取单个任务的总览卡片",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{task_id}/readiness": {
      "get": {
        "description": "Pre-flight check. ``checks`` lists blocking problems only.",
        "operationId": "get_task_readiness_api_tasks__task_id__readiness_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "任务就绪检查",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}